+++ CHANGELOG.md
@@ -9,7 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 
+* pkg: vscode: `update` now works on Windows
+
 * pkg: vscode: link settings in the correct macOS-specific place
+
+* pkg: vscode: use the correct settings path on Windows
 
 ## [0.3.1] - 2018-02-26